requirements.txt
flask[async]
semantic-kernel>=1.0.0
redis
python-dotenv